+++ VerzijaBezAppUser/RentApp/RentApp/Persistance/Repository/IBranchOfficeRepository.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 
 namespace RentApp.Persistance.Repository
 {
-    public interface IBranchOfficeRepository : IRepository<BranchOffice, int>
+    public interface IBranchOfficeRepository : IRepository<BranchOffice, long>
     {
         IEnumerable<BranchOffice> GetAll(int pageIndex, int pageSize);
     }
