+++ VerzijaBezAppUser/RentApp/RentApp/Persistance/Repository/IRatingRepository.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 
 namespace RentApp.Persistance.Repository
 {
-    public interface IRatingRepository : IRepository<Rating, int>
+    public interface IRatingRepository : IRepository<Rating, long>
     {
         IEnumerable<Rating> GetAll(int pageIndex, int pageSize);
     }
