+++ VerzijaBezAppUser/RentApp/RentApp/Persistance/Repository/IReservationRepository.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 
 namespace RentApp.Persistance.Repository
 {
-    public interface IReservationRepository : IRepository<Reservation, int>
+    public interface IReservationRepository : IRepository<Reservation, long>
     {
         IEnumerable<Reservation> GetAll(int pageIndex, int pageSize);
     }
