+++ VerzijaBezAppUser/RentApp/RentApp/Persistance/Repository/IServiceRepository.cs
@@ -3,8 +3,9 @@ using System.Collections.Generic;
 
 namespace RentApp.Persistance.Repository
 {
-    public interface IServiceRepository : IRepository<Service, int>
+    public interface IServiceRepository : IRepository<Service, long>
     {
         IEnumerable<Service> GetAll(int pageIndex, int pageSize);
+
     }
 }
