+++ VerzijaBezAppUser/RentApp/RentApp/Persistance/Repository/IVehicleRepository.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 
 namespace RentApp.Persistance.Repository
 {
-    public interface IVehicleRepository : IRepository<Vehicle, int>
+    public interface IVehicleRepository : IRepository<Vehicle, long>
     {
         IEnumerable<Vehicle> GetAll(int pageIndex, int pageSize);
     }
