+++ VerzijaBezAppUser/RentApp/RentApp/Persistance/Repository/IVehicleTypeRepository.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 
 namespace RentApp.Persistance.Repository
 {
-    public interface IVehicleTypeRepository : IRepository<VehicleType, int>
+    public interface IVehicleTypeRepository : IRepository<VehicleType, long>
     {
         IEnumerable<VehicleType> GetAll(int pageIndex, int pageSize);
     }
